+++ src/imode_gui/widgets/img.rs
@@ -5,15 +5,15 @@ use image::RgbaImage;
 
 impl Ui {
 
-    pub fn image(&mut self, id: TextureId, size: na::Vector2::<i32>){
-        let mut rect = Rect { x: 0, y: 0, w: size.x, h: size.y };
+    pub fn image(&mut self, id: TextureId, size: na::Vector2::<f32>){
+        let mut rect = Rect { x: 0, y: 0, w: size.x as i32, h: size.y as i32 };
 
         rect = self.layout_rect(rect);
 
         self.image_at(id, size, rect.x, rect.y);
     }
 
-    pub fn image_at(&mut self, id: TextureId, size: na::Vector2::<i32>, x: i32, y: i32) {
+    pub fn image_at(&mut self, id: TextureId, size: na::Vector2::<f32>, x: i32, y: i32) {
         self.drawer2D.render_img(id, x, y, size);
     }
